requirements.txt
fastapi
uvicorn
langchain
langgraph
python-dotenv
langchain-google-genai
jinja2